Add comment on SSL socket security
It turns out that decent security features are only available
starting with python 2.7.9.  As that is not yet widely available,
I cannot use it.  This means that the server will accept SSLv3
and RC4 without complaining.  I should blacklist these modes
and ciphers on the client-side, but still, it makes me sad.

I hope that a new version of python will be widely available soon.