Add pyVICBF implementation
Added the pyVICBF library (see github.com/malexmave/pyVICBF).
Would have loved to add it as a git submodule, but that is sadly
not possible.  And since I have not yet written an installer for
the library, I also can't put it into the requirements.txt, so
it'll be a hard copy for now, and I may or may not change it into
a pip-installable requirement at some later point.